requirements.txt
flask==2.3.3
flask-cors==4.0.0
onnxruntime==1.22.0
opencv-python-headless==4.10.0.84
pillow==10.4.0
numpy>=1.24.0
gunicorn==21.2.0